+++ src/react/mod.rs
@@ -6,10 +6,10 @@ use crate as bevy_cobweb;
 //module tree
 mod command_queue;
 mod commands;
-mod commands_ext;
 mod despawn_reader;
 mod entity_reaction_readers;
 mod event_readers;
+mod extensions;
 mod plugin;
 mod react_cache;
 mod react_commands;
@@ -21,14 +21,15 @@ mod reaction_triggers_impl;
 mod system_command_spawning;
 mod system_event_reader;
 mod utils;
+mod world_reactor;
 
 //API exports
 pub(crate) use crate::react::command_queue::*;
 pub use crate::react::commands::*;
-pub use crate::react::commands_ext::*;
 pub use crate::react::despawn_reader::*;
 pub use crate::react::entity_reaction_readers::*;
 pub use crate::react::event_readers::*;
+pub use crate::react::extensions::*;
 pub use crate::react::plugin::*;
 pub(crate) use crate::react::react_cache::*;
 pub use crate::react::react_commands::*;
@@ -40,3 +41,4 @@ pub use crate::react::reaction_triggers_impl::*;
 pub use crate::react::system_command_spawning::*;
 pub use crate::react::system_event_reader::*;
 pub use crate::react::utils::*;
+pub use crate::react::world_reactor::*;